requirements.txt
discord.py==2.3.2
flask==3.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-auth==2.23.4
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
datetime
pytz==2023.3
